[app] s/device_list_widget.dart/device_list.dart/
